Remove warnings.
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@5035 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ cm_pbio.c
@@ -390,7 +390,7 @@ int cond;
     vec[0].iov_len = sizeof(msg);
     vec[1].iov_base = format_ID;
     vec[1].iov_len = format_ID_length;
-    CMtrace_out(conn->cm, CMLowLevelVerbose, "CMpbio send format request - total %ld bytes in writev", format_ID_length + sizeof(msg));
+    CMtrace_out(conn->cm, CMLowLevelVerbose, "CMpbio send format request - total %ld bytes in writev", (long)format_ID_length + sizeof(msg));
     actual = conn->trans->writev_func(&CMstatic_trans_svcs, 
 				      conn->transport_data, 
 				      &vec[0], 2);
@@ -428,7 +428,7 @@ int cond;
     vec[0].iov_len = sizeof(msg);
     vec[1].iov_base = format_body_rep;
     vec[1].iov_len = body_len;
-    CMtrace_out(conn->cm, CMLowLevelVerbose, "CMpbio send format response - total %ld bytes in writev", body_len + sizeof(msg));
+    CMtrace_out(conn->cm, CMLowLevelVerbose, "CMpbio send format response - total %ld bytes in writev", (long)body_len + sizeof(msg));
     actual = conn->trans->writev_func(&CMstatic_trans_svcs, 
 				      conn->transport_data, 
 				      &vec[0], 2);
@@ -470,7 +470,7 @@ CMConnection conn;
     vec[1].iov_len = id_len;
     vec[2].iov_base = format_body_rep;
     vec[2].iov_len = body_len;
-    CMtrace_out(conn->cm, CMLowLevelVerbose, "CMpbio send format preload - total %ld bytes in writev", body_len + id_len + sizeof(msg));
+    CMtrace_out(conn->cm, CMLowLevelVerbose, "CMpbio send format preload - total %ld bytes in writev", (long)body_len + id_len + sizeof(msg));
     actual = conn->trans->writev_func(&CMstatic_trans_svcs, 
 				      conn->transport_data, 
 				      &vec[0], 3);
@@ -606,7 +606,7 @@ CM_pbio_query(CMConnection conn, CMTransport trans, char *buffer, int length)
 						    ((char*)&tmp_msg) + 8, 
 						    (int)sizeof(tmp_msg) - 8, 0);
 	    CMtrace_out(conn->cm, CMLowLevelVerbose, "CMpbio reading %ld msg bytes",
-			sizeof(tmp_msg) - 8);
+			(long)sizeof(tmp_msg) - 8);
 	    if (actual != (sizeof(tmp_msg) - 8)) {
 		CMtrace_out(conn->cm, CMLowLevelVerbose, 
 			    "CMdata read failed, actual %d", actual);
